rename MockEMRAndS3TestCase to MockBotoTestCase, move to tests.mockboto
